ref: Extract symbolicator-sources as its own workspace crate
