+++ crates/symbolicator-sources/src/lib.rs
@@ -1,0 +1,9 @@
+mod filetype;
+mod paths;
+mod sources;
+mod types;
+
+pub use filetype::*;
+pub use paths::*;
+pub use sources::*;
+pub use types::*;
