+++ crates/symbolicator/src/main.rs
@@ -32,7 +32,6 @@ mod endpoints;
 mod logging;
 mod server;
 mod services;
-mod sources;
 mod types;
 mod utils;
