+++ crates/symbolicator/src/services/download/locations.rs
@@ -10,8 +10,9 @@ use anyhow::{Error, Result};
 use serde::{Deserialize, Serialize};
 use url::Url;
 
+use symbolicator_sources::SourceId;
+
 use crate::services::cacher::CacheKey;
-use crate::sources::SourceId;
 use crate::types::Scope;
 use crate::utils::sentry::ConfigureScope;
