+++ crates/symbolicator/src/services/symcaches/mod.rs
@@ -6,9 +6,11 @@ use std::time::Duration;
 
 use anyhow::Error;
 use futures::future::BoxFuture;
+use thiserror::Error;
+
 use symbolic::common::{Arch, ByteView};
 use symbolic::symcache::{self, SymCache, SymCacheConverter};
-use thiserror::Error;
+use symbolicator_sources::{FileType, ObjectId, ObjectType, SourceConfig};
 
 use crate::cache::{Cache, CacheStatus};
 use crate::services::bitcode::BitcodeService;
@@ -17,10 +19,7 @@ use crate::services::objects::{
     FindObject, FoundObject, ObjectError, ObjectHandle, ObjectMetaHandle, ObjectPurpose,
     ObjectsActor,
 };
-use crate::sources::{FileType, SourceConfig};
-use crate::types::{
-    AllObjectCandidates, ObjectFeatures, ObjectId, ObjectType, ObjectUseInfo, Scope,
-};
+use crate::types::{AllObjectCandidates, ObjectFeatures, ObjectUseInfo, Scope};
 use crate::utils::futures::{m, measure};
 use crate::utils::sentry::ConfigureScope;
 
@@ -478,10 +477,10 @@ mod tests {
     use crate::config::{CacheConfigs, Config};
     use crate::services::bitcode::BitcodeService;
     use crate::services::DownloadService;
-    use crate::sources::{
+    use crate::test::{self, fixture};
+    use symbolicator_sources::{
         CommonSourceConfig, DirectoryLayoutType, FilesystemSourceConfig, SourceConfig, SourceId,
     };
-    use crate::test::{self, fixture};
 
     /// Creates a `SymCacheActor` with the given cache directory
     /// and timeout for download cache misses.
