+++ crates/symbolicator/src/utils/mod.rs
@@ -4,5 +4,4 @@ pub mod futures;
 pub mod gcs;
 pub mod hex;
 pub mod http;
-pub mod paths;
 pub mod sentry;
